Better amp control
Introduce a class to encapsulate comms with the amp. Prefer 12V trigger
for power on/off vs IR power toggle.

The yamaha 12V trigger is a fucking pain in the ass. It requires the
source to be set to main direct for both triggering off and it is really
finicky about how quickly trigger on can be issued after trigger off. It
will ignore trigger on if it comes in, based on my experiments, less
than ~700ms after the trigger off.